requirements.txt
django
django-crispy-forms
django-bootstrap-modal-forms
djangorestframework
markdown
django-filter